Add files via upload
fixed the issue where a undefined object is in the array. Index 6, 15, & 24 are ad objects and not listed restaurants. 
Also, fixed some of the css so that the buttons align centered in a column fashion. 

working on making the data structure based on a hash table and function to prevent repeats and such.